Make compatible with Tcl 9
This means relying on the default mechanism to
find Tclsignal_Init in the shared library.
Otherwise on Tcl 9.0 it looks for 'tclsignal_Init'
(lowercase 't') and won't find it.
This is due to a breaking change that titlecase is not applied
to the prefix (https://www.tcl.tk/man/tcl9.0/TclCmd/load.html)
The change still works in Tcl 8 because the library is correctly named.